Solved comment in pr
