+++ src/main/java/salesforce/utils/strategy/MapPages.java
@@ -6,7 +6,7 @@
  * license agreement you entered into with Fundacion Jala
  */
 
-package salesforce.utlis.strategy;
+package salesforce.utils.strategy;
 
 import org.openqa.selenium.InvalidArgumentException;
 import salesforce.ui.pages.lightning.contracts.ContractsPage;
