+++ src/main/java/salesforce/utils/supplier/StringSupplier.java
@@ -6,11 +6,8 @@
  * license agreement you entered into with Fundacion Jala
  */
 
-package salesforce.utlis.strategy;
+package salesforce.utils.supplier;
 
-import java.util.List;
-import java.util.Map;
-
-public interface ICreatedFeature {
-    List<String> getValueField(Map<String, String> table);
+public interface StringSupplier {
+    String getAsString();
 }
